Move zzrm into tex2pdf
