--- zerozeroreadme/mypy.ini
@@ -1,34 +1,0 @@
-[mypy]
-mypy_path = zerozeroreadme
-
-#
-# Covered by --strict, with some turned off:
-#
-disallow_untyped_defs=True
-check_untyped_defs=True
-# currently an issue with sql alchemy
-disallow_subclassing_any=false
-# Need to experiment/think about this one:
-disallow_any_decorated=false
-warn_redundant_casts=True
-warn_return_any=True
-warn_unused_ignores=True
-# this seems to be at least somewhat non-functioning:
-warn_unused_configs=True
-#may be worth reconsidering this one:
-no_implicit_optional=True
-strict_optional=True
-
-#
-# Other: 
-#
-ignore_missing_imports=True
-follow_imports=silent
-
-#
-# Plugins:
-#
-plugins = pydantic.mypy
-
-[tex_inspection.*]
-disallow_untyped_calls=True
