--- zerozeroreadme/pyproject.toml
@@ -1,35 +1,0 @@
-[tool.poetry]
-name = "zerozeroreadme"
-description = "00README parser and tools"
-version = "0.0.1"
-authors = [ "Norbert Preining"  ]
-
-[tool.poetry.dependencies]
-python = "^3.11"
-ruamel-yaml = "^0.18.5"
-toml = "^0.10.2"
-tomli_w = "^1.0"
-preflight_parser = {git = "https://github.com/arXiv/submission-tools.git", subdirectory = "preflight_parser" }
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.3.3"
-mypy = "*"
-mypy-extensions = "*"
-pydantic = "==1.10.*"
-types-toml = "^0.10.8.20240310"
-
-[tool.ruff]
-# Extend the `pyproject.toml` from the toplevel dir
-extend = "../pyproject.toml"
-extend-exclude = ["tests"]
-
-[tool.black]
-line-length = 120
-
-[tool.pylint]
-max-line-length = 120
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
